tentando adicionar testes de compra
